LLVM JIT with Orc works, but is unfinished.
The main problem with it is that it appears that it will not recognize
functions added to the module after you’ve added it to the JIT.
Presumably because some sort of symbol connectivity pre-pass is done.
This restriction makes Orc no different from MCJIT as far as I can
tell. There may be some hidden way to do this that I am not aware of,
but it certainly isn’t exposed in the C API.

This make it hard to do reentrant parsing since I’d need 2 passes to
mark which functions I can safely put into the module that won’t
require #run.

I may just writing byte code at this point.